LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/jsurl-navigation-then-form-submit-expected.txt



FAIL Verifies that form submissions cancel javascript navigations to prevent duplicate load events. assert_equals: expected "/formaction.html" but got "blank"
